Time: 9 ms (39.73%), Space: 45.8 MB (41.44%) - LeetHub

+++ 1504-count-submatrices-with-all-ones/1504-count-submatrices-with-all-ones.java
@@ -4,7 +4,7 @@ class Solution {
 
         int m = mat.length, n = mat[0].length;
         int[] heights = new int[n];
-        long ans = 0;
+        int ans = 0;
 
         for (int r = 0; r < m; r++) {
             // 1) Build histogram for this row
@@ -38,7 +38,7 @@ class Solution {
                 ans += sum;
             }
         }
-        return (int)ans;
+        return ans;
         
     }
 }
